// src/rust-port/client/model.ts
import { AttestationType, DeviceType } from '../../api/models';
import type { AuthKey, Udid } from '../protocol';

export interface RequestBase {
  carrier: string;
  device: number;
  device_id: string;
  device_name: string;
  dmm_onetime_token: null;
  dmm_viewer_id: null;
  graphics_device_name: string;
  ip_address: string;
  keychain: number;
  locale: string;
  platform_os_version: string;
  viewer_id: number;
  steam_id: string | null;
  steam_session_ticket: string | null;
}

export interface DataHeaders {
  viewer_id: number;
  sid: string;
  servertime: number;
  result_code: number;
  [k: string]: unknown;
}

export interface UmaResponse<T> {
  data_headers: DataHeaders;
  data?: T;
}

export interface SignupRequest {
  error_code: number;
  error_message: string;
  attestation_type: number;
  optin_user_birth: number;
  dma_state: number;
  country: string;
  credential: string;
}

export interface SignupData {
  viewer_id: number;
  auth_key: string;
}

export interface StartSessionRequest {
  attestation_type: number;
  device_token: string | null;
}

export interface StartSessionResponse {
  attest: boolean;
  is_tutorial: boolean;
  nonce: string;
  resource_version: string;
  terms_updated: boolean;
}

export interface ClientConfig {
  udid?: Udid;
  authKey?: AuthKey;
  base?: RequestBase;
}

export type AuthMode =
  | {
      kind: 'steam';
      username: string;
      password: string;
    }
  | {
      kind: 'custom';
      deviceType: DeviceType;
      attestationType: AttestationType;
    };